+++ tests/browser/query_all.test.js
@@ -45,14 +45,55 @@ describe("Query All", function() {
         assert(elements[0]);
     });
 
-    it("QueryAll Sub Element Not Valid Parent", async() => {
-        await utils.assertThrowsAsync(async() => {
-            await browser.queryAll(".container", "b");
-        }, "Error: [queryAll] Invalid parent element for queryAll");
-    });
-
     it("QueryAll Element Not Found", async() => {
         const elements = await browser.queryAll(".not-element");
         assert.strictEqual(elements.length, 0);
+    });
+
+    it("Query Invalid Selector", async() => {
+        await utils.assertThrowsAsync(async() => {
+            await browser.query(10);
+        }, `Error: [query] Invalid selector.`);
+    });
+
+    it("QueryAll Invalid Selector", async() => {
+        await utils.assertThrowsAsync(async() => {
+            await browser.queryAll(10);
+        }, `Error: [queryAll] Invalid selector.`);
+    });
+
+    it("Query Invalid Selector With Optional Selector", async() => {
+        await utils.assertThrowsAsync(async() => {
+            await browser.query(10, ".p");
+        }, `Error: [query] Invalid selector.`);
+    });
+
+    it("QueryAll Invalid Selector With Optional Selector", async() => {
+        await utils.assertThrowsAsync(async() => {
+            await browser.queryAll(10, ".p");
+        }, `Error: [queryAll] Invalid selector.`);
+    });
+
+    it("QueryAll With XPath", async() => {
+        const elements = await browser.queryAll('//p[contains(text(),"paragraph")]');
+        assert.strictEqual(elements.length, 2);
+    });
+
+    it("QueryAll With XPath Children", async() => {
+        const elements = await browser.queryAll("body", '//p[contains(text(),"paragraph")]');
+        assert.strictEqual(elements.length, 2);
+    });
+
+    it("QueryAll With XPath Children With One Element Only", async() => {
+        const elements = await browser.queryAll(".container", '//p[contains(text(),"paragraph")]');
+        assert.strictEqual(elements.length, 1);
+    });
+
+    it("QueryAll With Multiple Parents", async() => {
+        await browser.open(configUrls.nestedElements);
+        const elements = await browser.queryAll(".container", "p");
+        assert.strictEqual(elements.length, 2);
+        await browser.assert.text(elements[0], "My first paragraph");
+        await browser.assert.text(elements[1], "My second paragraph");
     });
 });